pyproject.toml
[project]
name = "ETF-Investment-Simulator"
version = "1.0.0"
description = "A Flask-based web application for simulating fixed-amount investment strategies"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["flask", "investment", "etf", "simulator", "financial"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: Flask",
    "Development Status :: 3 - Alpha",
]
dependencies = [
    "Flask==3.0.0",
    "numpy==1.24.3",
    "matplotlib==3.7.1",
    "python-dotenv==1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ETF"
Repository = "https://github.com/yourusername/ETF.git"
Issues = "https://github.com/yourusername/ETF/issues"

[build-system]
requires = ["setuptools>=40.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
# uv configuration

[tool.uv.sources]
# Optional: pin sources for specific packages

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100
